app/requirements.txt
fastapi
uvicorn
sqlalchemy
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-prometheus
psycopg2-binary
